mejora la interface de entrega

+++ base/funciones/conta.ft_entrega_det_ime.sql
@@ -1,8 +1,13 @@
-CREATE OR REPLACE FUNCTION "conta"."ft_entrega_det_ime" (	
-				p_administrador integer, p_id_usuario integer, p_tabla character varying, p_transaccion character varying)
-RETURNS character varying AS
-$BODY$
+--------------- SQL ---------------
 
+CREATE OR REPLACE FUNCTION conta.ft_entrega_det_ime (
+  p_administrador integer,
+  p_id_usuario integer,
+  p_tabla varchar,
+  p_transaccion varchar
+)
+RETURNS varchar AS
+$body$
 /**************************************************************************
  SISTEMA:		Sistema de Contabilidad
  FUNCION: 		conta.ft_entrega_det_ime
@@ -27,6 +32,8 @@ DECLARE
 	v_nombre_funcion        text;
 	v_mensaje_error         text;
 	v_id_entrega_det	integer;
+    v_registros			record;
+    v_c31				varchar;
 			    
 BEGIN
 
@@ -43,31 +50,60 @@ BEGIN
 	if(p_transaccion='CONTA_END_INS')then
 					
         begin
-        	--Sentencia de la insercion
+        	
+            --verificar que no este asociado a otra entrega
+            
+            
+             v_c31 = null;
+             select 
+                c.c31
+              into 
+                v_c31
+             from conta.tint_comprobante c 
+             where c.id_int_comprobante =  v_parametros.id_int_comprobante;
+                 
+             if v_c31 is not null  and trim(v_c31) != ''   then                 
+                raise exception 'El comprobantes (id: %)  ya se encuentra relacionado con la entrega o C31: %',  v_parametros.id_int_comprobante ,v_c31;
+             end if;
+            
+            IF v_parametros.id_entrega is null THEN
+               raise exception 'El id de la entrega no puede ser nulo';
+            END IF;
+            
+            --Sentencia de la insercion
         	insert into conta.tentrega_det(
-			estado_reg,
-			id_int_comprobante,
-			id_entrega,
-			id_usuario_reg,
-			fecha_reg,
-			usuario_ai,
-			id_usuario_ai,
-			id_usuario_mod,
-			fecha_mod
+              estado_reg,
+              id_int_comprobante,
+              id_entrega,
+              id_usuario_reg,
+              fecha_reg,
+              usuario_ai,
+              id_usuario_ai,
+              id_usuario_mod,
+              fecha_mod
           	) values(
-			'activo',
-			v_parametros.id_int_comprobante,
-			v_parametros.id_entrega,
-			p_id_usuario,
-			now(),
-			v_parametros._nombre_usuario_ai,
-			v_parametros._id_usuario_ai,
-			null,
+              'activo',
+              v_parametros.id_int_comprobante,
+              v_parametros.id_entrega,
+              p_id_usuario,
+              now(),
+              v_parametros._nombre_usuario_ai,
+              v_parametros._id_usuario_ai,
+              null,
 			null
 							
 			
 			
 			)RETURNING id_entrega_det into v_id_entrega_det;
+            
+            
+             --  temporalmente marca el cbte relacionado a la entrega
+             update conta.tint_comprobante  set
+                 c31 =  'ENT ID:'||v_parametros.id_entrega::varchar,
+                 fecha_c31 = now()
+              where id_int_comprobante = v_parametros.id_int_comprobante;
+            
+            --marca el cbte como relacionado
 			
 			--Definicion de la respuesta
 			v_resp = pxp.f_agrega_clave(v_resp,'mensaje','Entrega Detalle almacenado(a) con exito (id_entrega_det'||v_id_entrega_det||')'); 
@@ -78,35 +114,7 @@ BEGIN
 
 		end;
 
-	/*********************************    
- 	#TRANSACCION:  'CONTA_END_MOD'
- 	#DESCRIPCION:	Modificacion de registros
- 	#AUTOR:		admin	
- 	#FECHA:		17-11-2016 19:50:46
-	***********************************/
-
-	elsif(p_transaccion='CONTA_END_MOD')then
-
-		begin
-			--Sentencia de la modificacion
-			update conta.tentrega_det set
-			id_int_comprobante = v_parametros.id_int_comprobante,
-			id_entrega = v_parametros.id_entrega,
-			id_usuario_mod = p_id_usuario,
-			fecha_mod = now(),
-			id_usuario_ai = v_parametros._id_usuario_ai,
-			usuario_ai = v_parametros._nombre_usuario_ai
-			where id_entrega_det=v_parametros.id_entrega_det;
-               
-			--Definicion de la respuesta
-            v_resp = pxp.f_agrega_clave(v_resp,'mensaje','Entrega Detalle modificado(a)'); 
-            v_resp = pxp.f_agrega_clave(v_resp,'id_entrega_det',v_parametros.id_entrega_det::varchar);
-               
-            --Devuelve la respuesta
-            return v_resp;
-            
-		end;
-
+	
 	/*********************************    
  	#TRANSACCION:  'CONTA_END_ELI'
  	#DESCRIPCION:	Eliminacion de registros
@@ -117,6 +125,29 @@ BEGIN
 	elsif(p_transaccion='CONTA_END_ELI')then
 
 		begin
+        
+       -- raise exception 'id_entrega_det %', v_parametros.id_entrega_det;
+        
+           select 
+               e.estado,
+               e.id_entrega,
+               ed.id_int_comprobante
+           into
+              v_registros
+           from conta.tentrega_det ed
+           inner join conta.tentrega e on e.id_entrega = ed.id_entrega
+           where ed.id_entrega_det = v_parametros.id_entrega_det;
+        
+          if v_registros.estado != 'borrador' then
+             raise exception 'Solo puede quitar las relaciones en entregas borrador';
+          end if;
+        
+            --desmarcar el cbte como no relacionado
+             update conta.tint_comprobante  set
+               c31 = NULL,
+               fecha_c31 = NULL
+             where id_int_comprobante=v_registros.id_int_comprobante;
+             
 			--Sentencia de la eliminacion
 			delete from conta.tentrega_det
             where id_entrega_det=v_parametros.id_entrega_det;
@@ -146,7 +177,9 @@ EXCEPTION
 		raise exception '%',v_resp;
 				        
 END;
-$BODY$
-LANGUAGE 'plpgsql' VOLATILE
+$body$
+LANGUAGE 'plpgsql'
+VOLATILE
+CALLED ON NULL INPUT
+SECURITY INVOKER
 COST 100;
-ALTER FUNCTION "conta"."ft_entrega_det_ime"(integer, integer, character varying, character varying) OWNER TO postgres;
